Crate linked_vector

Structs

Cursor

A cursor which can only read the elements of the list.

CursorMut

A cursor which can read and write the elements of the list.

HNode

A handle to a node within a LinkedVector. Internally, it holds an index into the vector holding the LinkedVector’s nodes.

Handles

An iterator over the elements of a LinkedVector. Yields the handles of each element.

IntoIter

The consuming iterator class of LinkedVector. Yields owned elements of the vector.

Iter

The basic iterator class of LinkedVector. Yields references to the elements of the vector.

IterMut

The basic iterator class of LinkedVector. Yields mutable references to the elements of the vector.

LinkedVector

A doubly-linked list that uses handles to refer to elements that exist within a vector. This allows for O(1) insertion and removal of elements from the list, and O(1) access to elements by handle.

Traits

CursorBase

A cursor is a position within a linked vector. It can be used to traverse the list in either direction, and to access the element at the current position.