Struct linked_vector::CursorMut

pub struct CursorMut<'a, T> { /* private fields */ }

A cursor which can read and write the elements of the list. CursorMut supports a remove() method which removes the element at the current position of the cursor. This feature is turned on with the "cursor-remove" feature, see usage notes.

Implementations

impl<'a, T> CursorMut<'a, T>

pub fn is_empty(&self) -> bool

Returns true if the vector the cursor is attached to is empty. Since a mutable cursor can remove items, this is provided as a means to avoid panics if the cursor is being used to remove items. If the underlying vector is empty, other operations may panic.

pub fn get_mut(&mut self) -> Option<&mut T>

Returns a mutable reference to the element at the cursor’s current position, or None if the underlying vector is empty. If the optionless-accessors feature is enabled, this will return a reference directly. This feature is disabled by default, see usage notes.

pub fn insert(&mut self, value: T) -> HNode

Inserts a new element at the cursor’s current position. The cursor will be moved to the new element. Returns the handle of the new element.

pub fn insert_after(&mut self, value: T) -> HNode

Inserts a new element after the cursor’s current position. The cursor will still be at the same position. Returns the handle of the new element.